Delay valve relay state update
The controller will wait 5 minutes on the production before switching
the state of a given valve. If it will get a new value within that time
period, it will wait another 5 minutes. This way, the relay won't be
changed every minute because the temperature is oscilating near the
expected value.